@@ -3,7 +3,7 @@ source 'https://rubygems.org'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.5'
+gem 'rails', '5.0.0.1'
 # Use postgresql as the database for Active Record
 gem 'pg', '~> 0.15'
 # Use SCSS for stylesheets
@@ -28,7 +28,7 @@ gem 'sidekiq'
 gem 'sidetiq'
 gem 'sidekiq-unique-jobs'
 gem 'sidekiq-failures'
-gem 'sinatra'
+gem 'sinatra', '2.0.0.beta2'
 gem 'aws-sdk', '~> 2.0'
 gem 'annotate'
 gem 'rails_12factor', group: :production
@@ -77,7 +77,7 @@ end
 
 group :test do
   gem 'database_cleaner'
-  gem 'rspec-rails', '~> 3.3.3'
+  gem 'rspec-rails'
   gem 'rspec-its'
   gem 'rspec-collection_matchers'
   gem 'spring-commands-rspec', require: false
