@@ -1,7 +1,7 @@
 class ApplicationController < ActionController::Base
   force_ssl if Rails.env.production?
   protect_from_forgery with: :exception
-  after_filter :flash_to_http_header
+  after_action :flash_to_http_header
 
   private
 
