@@ -11,13 +11,13 @@ describe Api::V1::BaseController, type: :request do
     context 'creating tokens' do
       it 'creates a devise api auth token' do
         expect(user.tokens).to be_empty
-        post '/api/v1/auth/sign_in', sign_in_data.to_json, base_headers
+        post '/api/v1/auth/sign_in', params: sign_in_data.to_json, headers: base_headers
         expect(user.reload.tokens).to_not be_empty
       end
     end
 
     context 'the response' do
-      before        { post '/api/v1/auth/sign_in', sign_in_data.to_json, base_headers }
+      before        { post '/api/v1/auth/sign_in', params: sign_in_data.to_json, headers: base_headers }
       subject       { response }
       its(:headers) { is_expected.to have_key('access-token') }
       its(:headers) { is_expected.to have_key('client') }
@@ -28,7 +28,7 @@ describe Api::V1::BaseController, type: :request do
 
   describe 'DELETE /api/v1/auth/sign_out' do
     before do
-      post '/api/v1/auth/sign_in', sign_in_data.to_json, base_headers
+      post '/api/v1/auth/sign_in', params: sign_in_data.to_json, headers: base_headers
       @auth_creds = response.headers
                      .slice('access-token', 'client', 'uid', 'expiry')
                      .merge('token-type' => 'Bearer')
@@ -36,7 +36,7 @@ describe Api::V1::BaseController, type: :request do
 
     it 'removes the token' do
       expect {
-        delete '/api/v1/auth/sign_out', {}, base_headers.merge(@auth_creds)
+        delete '/api/v1/auth/sign_out', headers: base_headers.merge(@auth_creds)
       }.to change { user.reload.tokens }
     end
   end
@@ -44,7 +44,7 @@ describe Api::V1::BaseController, type: :request do
   describe 'GET test' do
     context 'an authorized user' do
       before do
-        post '/api/v1/auth/sign_in', sign_in_data.to_json, base_headers
+        post '/api/v1/auth/sign_in', params: sign_in_data.to_json, headers: base_headers
         @auth_creds = response.headers
                        .slice('access-token', 'client', 'uid', 'expiry')
                        .merge('token-type' => 'Bearer')
@@ -52,13 +52,13 @@ describe Api::V1::BaseController, type: :request do
 
       context 'no server error' do
         it 'can successfully access a protected endpoint if authorized' do
-          get '/api/v1/test', {}, base_headers.merge(@auth_creds)
+          get '/api/v1/test', headers: base_headers.merge(@auth_creds)
           expect(response).to be_ok
           expect(response.body).to eq('ok')
         end
 
         context 'the headers of the response' do
-          before        { get '/api/v1/test', {}, base_headers.merge(@auth_creds) }
+          before        { get '/api/v1/test', headers: base_headers.merge(@auth_creds) }
           subject       { response }
           its(:headers) { is_expected.to have_key('access-token') }
           its(:headers) { is_expected.to have_key('client') }
@@ -76,7 +76,7 @@ describe Api::V1::BaseController, type: :request do
           end
 
           it 'does something' do
-            get '/api/v1/test', {}, base_headers.merge(@auth_creds)
+            get '/api/v1/test', headers: base_headers.merge(@auth_creds)
             expect(response).to be_unauthorized
           end
         end
@@ -85,7 +85,7 @@ describe Api::V1::BaseController, type: :request do
 
     context 'a non authorized user' do
       it 'responds with an unauthorized status' do
-        get '/api/v1/test', {}, base_headers
+        get '/api/v1/test', headers: base_headers
         expect(response).to be_unauthorized
       end
     end
